add javadocs documentation to files in the search package and move publishing related activities to the publish package
