+++ WiseTrack/app/src/main/AndroidManifest.xml
@@ -10,10 +10,10 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.WiseTrack">
 
-        <activity android:name=".PublishExperimentActivity4"/>
-        <activity android:name=".PublishExperimentActivity3" />
-        <activity android:name=".PublishExperimentActivity2" />
-        <activity android:name=".PublishExperimentActivity" />
+        <activity android:name=".publish.PublishExperimentActivity4"/>
+        <activity android:name=".publish.PublishExperimentActivity3" />
+        <activity android:name=".publish.PublishExperimentActivity2" />
+        <activity android:name=".publish.PublishExperimentActivity" />
         <activity android:name=".search.SearchActivity" />
         <activity android:name=".MainMenuActivity"/>
