+++ WiseTrack/app/src/main/java/com/faanggang/wisetrack/publish/PublishExperimentActivity.java
@@ -1,4 +1,4 @@
-package com.faanggang.wisetrack;
+package com.faanggang.wisetrack.publish;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -7,6 +7,8 @@ import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
+
+import com.faanggang.wisetrack.R;
 
 public class PublishExperimentActivity extends AppCompatActivity {
