+++ WiseTrack/app/src/main/java/com/faanggang/wisetrack/publish/PublishExperimentActivity3.java
@@ -1,4 +1,4 @@
-package com.faanggang.wisetrack;
+package com.faanggang.wisetrack.publish;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -6,6 +6,8 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
+
+import com.faanggang.wisetrack.R;
 
 import java.util.EmptyStackException;
