+++ WiseTrack/app/src/main/java/com/faanggang/wisetrack/publish/PublishExperimentActivity4.java
@@ -1,4 +1,4 @@
-package com.faanggang.wisetrack;
+package com.faanggang.wisetrack.publish;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -10,6 +10,8 @@ import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
+import com.faanggang.wisetrack.MainActivity;
+import com.faanggang.wisetrack.R;
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.firebase.Timestamp;
