+++ WiseTrack/app/src/main/res/layout/activity_publish_experiment.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PublishExperimentActivity">
+    tools:context=".publish.PublishExperimentActivity">
 
     <EditText
         android:id="@+id/name_input"
