+++ WiseTrack/app/src/main/res/layout/activity_publish_experiment2.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PublishExperimentActivity2">
+    tools:context=".publish.PublishExperimentActivity2">
 
     <TextView
         android:id="@+id/choose_a_crowd_source_test"
