+++ WiseTrack/app/src/main/res/layout/activity_publish_experiment3.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PublishExperimentActivity3">
+    tools:context=".publish.PublishExperimentActivity3">
 
     <TextView
         android:id="@+id/require_geolocation"
