+++ WiseTrack/app/src/main/res/layout/activity_publish_experiment4.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PublishExperimentActivity4">
+    tools:context=".publish.PublishExperimentActivity4">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/scrollView3"
